The backend C library now consistently uses ltfatInt (ptrdiff_t or int) for array lengths. fftreal moved to the backend. dct and dst moved to the backend. filterbank routines moved to the backend. framematrix is now deprecated, and was replaced by frsynmatrix, which better reflect what the function does.

+++ mex/comp_pchirp.c
@@ -6,9 +6,9 @@
 
 #define PI 3.1415926535897932384626433832795
 
-static inline long positiverem_long(long a,long b)
+static inline long long positiverem_long(long long a,long long b)
 {
-  const long c = a%b;
+  const long long c = a%b;
   return(c<0 ? c+b : c);
 }
 
@@ -18,20 +18,20 @@ static inline long positiverem_long(long a,long b)
 void mexFunction( int nlhs, mxArray *plhs[], 
 		  int nrhs, const mxArray *prhs[] )
 { 
-   const long L=(long)mxGetScalar(prhs[0]);
-   const long n=(long)mxGetScalar(prhs[1]);
+   const long long L=(long long) mxGetScalar(prhs[0]);
+   const long long n=(long long) mxGetScalar(prhs[1]);
 
    plhs[0] = mxCreateDoubleMatrix(L, 1, mxCOMPLEX);
    double *gr = mxGetPr(plhs[0]);
    double *gi = mxGetPi(plhs[0]);
 
 
-   const long LL=2*L;
-   const long Lponen=positiverem_long((L+1)*n,LL);
+   const long long LL=2*L;
+   const long long Lponen=positiverem_long((L+1)*n,LL);
    
-   for (long m=0;m<L;m++)
+   for (long long m=0;m<L;m++)
    {
-      const long idx = positiverem_long(
+      const long long idx = positiverem_long(
    	 positiverem_long(Lponen*m,LL)*m,LL);
 
       gr[m] = cos(PI*idx/L);
